+++ common/src/main/java/net/mehvahdjukaar/moonlight/api/item/IItemDecoratorRenderer.java
@@ -1,6 +1,5 @@
 package net.mehvahdjukaar.moonlight.api.item;
 
-import com.mojang.blaze3d.vertex.PoseStack;
 import net.minecraft.client.gui.Font;
 import net.minecraft.client.gui.GuiGraphics;
 import net.minecraft.world.item.ItemStack;