+++ common/src/main/java/net/mehvahdjukaar/moonlight/core/Moonlight.java
@@ -1,9 +1,9 @@
 package net.mehvahdjukaar.moonlight.core;
 
-import net.mehvahdjukaar.moonlight.api.client.util.TextUtil;
 import net.mehvahdjukaar.moonlight.api.events.IDropItemOnDeathEvent;
 import net.mehvahdjukaar.moonlight.api.events.MoonlightEventsHelper;
 import net.mehvahdjukaar.moonlight.api.fluids.SoftFluidRegistry;
+import net.mehvahdjukaar.moonlight.api.integration.CompatWoodTypes;
 import net.mehvahdjukaar.moonlight.api.map.MapDecorationRegistry;
 import net.mehvahdjukaar.moonlight.api.misc.EventCalled;
 import net.mehvahdjukaar.moonlight.api.misc.RegistryAccessJsonReloadListener;
@@ -13,8 +13,8 @@ import net.mehvahdjukaar.moonlight.api.resources.pack.DynamicResourcePack;
 import net.mehvahdjukaar.moonlight.api.set.leaves.LeavesTypeRegistry;
 import net.mehvahdjukaar.moonlight.api.set.wood.WoodTypeRegistry;
 import net.mehvahdjukaar.moonlight.core.criteria_triggers.ModCriteriaTriggers;
-import net.mehvahdjukaar.moonlight.api.integration.CompatWoodTypes;
 import net.mehvahdjukaar.moonlight.core.loot_pool_entries.ModLootPoolEntries;
+import net.mehvahdjukaar.moonlight.api.item.additional_placements.BlockPlacerItem;
 import net.mehvahdjukaar.moonlight.core.misc.CaveFilter;
 import net.mehvahdjukaar.moonlight.core.misc.VillagerAIInternal;
 import net.mehvahdjukaar.moonlight.core.network.ModMessages;
@@ -25,9 +25,11 @@ import net.minecraft.core.registries.Registries;
 import net.minecraft.resources.ResourceLocation;
 import net.minecraft.tags.TagKey;
 import net.minecraft.world.entity.player.Player;
+import net.minecraft.world.item.Item;
 import net.minecraft.world.item.ItemStack;
 import net.minecraft.world.level.GameRules;
 import net.minecraft.world.level.block.Block;
+import net.minecraft.world.level.block.Blocks;
 import net.minecraft.world.level.levelgen.placement.PlacementModifierType;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
@@ -45,6 +47,10 @@ public class Moonlight {
 
     public static final Supplier<PlacementModifierType<CaveFilter>> CAVE_MODIFIER = RegHelper.register(
             res("below_heightmaps"), CaveFilter.Type::new, Registries.PLACEMENT_MODIFIER_TYPE);
+
+    public static final Supplier<BlockPlacerItem> BLOCK_PLACER = RegHelper.registerItem(
+            res("placeable_item"), () -> new BlockPlacerItem(
+                    Blocks.VOID_AIR, new Item.Properties()));
 
     public static ResourceLocation res(String name) {
         return new ResourceLocation(MOD_ID, name);
@@ -96,10 +102,10 @@ public class Moonlight {
         DynamicResourcePack.clearAfterReload(false);
     }
 
-    public static void assertInitPhase(){
-        if(!PlatHelper.isInitializing()){
+    public static void assertInitPhase() {
+        if (!PlatHelper.isInitializing()) {
             //TODO: re add once all mods are updated
-            if(PlatHelper.isDev()) {
+            if (PlatHelper.isDev()) {
                 throw new AssertionError("Method has to be called during mod initialization phase");
             }
         }